pyproject.toml
[project]
name = "torch-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pillow",
    "torch",
    "torchvision"
]
